@@ -1,38 +1,66 @@
 {
-  "name": "express-example",
-  "version": "0.0.0",
-  "private": true,
-  "scripts": {
-    "build": "./node_modules/.bin/webpack",
-    "server": "node bin/www",
-    "start": "npm run build && npm run server"
+  "name": "TravelSecure-Mongo",
+  "version": "3.3.7+1",
+  "homepage": "https://sleepy-springs-57038.herokuapp.com/",
+  "author": "Omar Patel",
+  "license": {
+    "type": "MIT"
   },
-  "dependencies": {
-    "axios": "^0.15.2",
-    "bcryptjs": "^2.3.0",
-    "body-parser": "~1.13.2",
-    "browser-sync": "^2.17.5",
-    "cookie-parser": "~1.3.5",
-    "debug": "~2.2.0",
-    "express": "~4.13.1",
-    "express-handlebars": "^2.0.1",
-    "express-session": "^1.14.1",
-    "gulp-nodemon": "^2.2.1",
-    "jsonwebtoken": "^7.1.9",
-    "method-override": "^2.3.5",
-    "mongoose": "^4.6.5",
-    "morgan": "~1.6.1",
-    "react": "^15.4.0",
-    "react-dom": "^15.4.0",
-    "serve-favicon": "~2.3.0"
+  "scripts": {
+    "dev": "webpack & node server",
+    "start": "nodemon ./server/bin/www",
+    "build": "webpack"
   },
   "devDependencies": {
-    "babel-core": "^6.18.2",
-    "babel-loader": "^6.2.7",
-    "babel-preset-es2015": "^6.18.0",
-    "babel-preset-react": "^6.16.0",
-    "chai": "^3.5.0",
-    "mocha": "^3.1.2",
-    "webpack": "^1.13.3"
+    "babel-core": "^6.3.13",
+    "babel-loader": "^6.2.0",
+    "babel-plugin-transform-class-properties": "^6.24.1",
+    "babel-preset-es2015": "^6.3.13",
+    "babel-preset-react": "^6.3.13",
+    "bootstrap": "^3.3.7",
+    "browser-sync": "^2.13.0",
+    "chai": "^4.0.0",
+    "css-loader": "^0.28.4",
+    "file-loader": "^0.11.2",
+    "font-awesome": "^4.6.3",
+    "jquery": "^1.11.3",
+    "redux-devtools": "^3.4.0",
+    "sinon": "^2.3.2",
+    "sinon-mongoose": "^2.0.2",
+    "style-loader": "^0.18.2",
+    "url-loader": "^0.5.9",
+    "webpack": "^1.13.1"
+  },
+  "plugins": [
+    "transform-class-properties"
+  ],
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/OSP123/TravelSecure-Mongo.git"
+  },
+  "dependencies": {
+    "axios": "^0.12.0",
+    "bcrypt-nodejs": "0.0.3",
+    "body-parser": "^1.17.1",
+    "connect-flash": "^0.1.1",
+    "cookie-parser": "^1.4.3",
+    "express": "^4.15.2",
+    "express-handlebars": "^3.0.0",
+    "express-session": "^1.15.2",
+    "jsonwebtoken": "^7.4.1",
+    "method-override": "^2.3.8",
+    "mongoose": "^4.10.0",
+    "morgan": "^1.8.1",
+    "nodemon": "^1.11.0",
+    "passport": "^0.3.2",
+    "passport-local": "^1.0.0",
+    "react": "^15.6.1",
+    "react-dom": "^15.6.1",
+    "react-redux": "^5.0.5",
+    "react-router-dom": "^4.1.1",
+    "redux": "^3.7.2"
+  },
+  "engines": {
+    "node": "7.8.0"
   }
 }
