@@ -1,26 +1,33 @@
 module.exports = {
-  
-  // This code will be compiled 
+
+  // This is the entry point or start of our react applicaton
   entry: "./app/app.js",
 
-  // Then output into this file
+  // The plain compiled JavaScript will be output into this file
   output: {
     filename: "public/bundle.js"
   },
 
-  // This will be what we do
+  // This section desribes the transformations we will perform
   module: {
     loaders: [
       {
+        // Only working with files that in in a .js or .jsx extension
         test: /\.jsx?$/,
-        exclude: /(node_modules|bower_components)/,
-        loader: 'babel',
+        // Webpack will only process files in our app folder. This avoids processing
+        // node modules and server files unnecessarily
+        include: /app/,
+        loader: "babel",
         query: {
-          // These are the specific transformations we'll be using. 
-          presets: ['react', 'es2015']
+          // These are the specific transformations we'll be using.
+          presets: ["react", "es2015"]
         }
-      }
+      },
+      {test: /\.css$/, loaders: ['style', 'css']},
+      {test: /\.(png|jpg|jpeg)$/, loader: 'url-loader'}
     ]
-  }
-
-}
+  },
+  // This lets us debug our react code in chrome dev tools. Errors will have lines and file names
+  // Without this the console says all errors are coming from just coming from bundle.js
+  devtool: "eval-source-map"
+};
